Export Logic cleanup first batch
Export was managed by many classes in different places due the nature of
a struct and all public variables.
This logic is now moved to class TExportSpecification with all private
data members and a first start of checking what data is set when and by
which object.

Second part will be to clear up ownership of data and remove code
duplication that started to happen along the way long long time ago.

If time allows a third (or in the second batch) the rendering will be
offloaded via the Qt Concurrent threading logic and not by inheriting
QThread. This will allow rendering to be halted and resumed and stopped
in a more sophisticated way then it is done now.

Since libsndfile now supports all the formats we really need (except for
WavPack) the traverso audiofileio library will be simplified by removing
the burden of maintaining the different audio file formats ourselves
(FLAC, Vorbis, MP3) and move it into libtraversocore