DiskIO: don't run on event loop but block on semaphore
freewheeling is dumping lots of events on the event queue and this
causes delayed events to be posted to the DiskIO event loop when it's
very busy (freewheeling).
Added event queue to DiskIO itself that is filled by Sheet:process()
instead audiodevice finishedOneCycle().
The benefit is that if DiskIO do_work() is very busy, next time it runs
it'll throw away all the added evens to it's queue since we don't need
them, but we now can calculate the amount of samples being processed by
the audio thread since our last round of do_work()