Freewheeling: start/stop drivers when freewheeling stops/starts
For this to work the easiest route is that all Drivers are blocking in
their write() function. PortAudio also can be used in blocking behavior
so as a test PortAudio driver now blocks instead of being based on a
callback from PortAudio.
PulseAudio/Alsa seem to behave nicely but PortAudio Driver not so much,
sometimes freewheeling won't start probably cause the audiothread is
blocked and waiting on Pa_WriteStream().
PortAudio needs some thourough review anyways but we need it for
Windows...