Code cleanup on Gain, ShortCut system
InputEventDispatcher/shortcut system:
Functions that are always save to dispatch from TInputEventDispatcher:
mark those with DISPATCH_RULE_IS_ALWAYS in front of the declaration.
This is compiled into a tag in the QMetaObject and can then be used to
identify if the input event dispatcher can call that shortcut function
during hold action or not instead of using a boolean flag when
registering the shortcut function.

Gain:
* Renamed to GainCommand and since setting gain is now done exclusively
on TAudioProcessingNodes, make the stored pointers explicitly
TAudioProcessingNode instead of generic QObjects and then having to use
QMetaObject::invoke(args)